Move GHC dependent files to subfolder for package
