+++ src/HIE/Bios.hs
@@ -16,5 +16,5 @@ module HIE.Bios (
 
 import HIE.Bios.Cradle
 import HIE.Bios.Types
-import HIE.Bios.GHCApi
-import HIE.Bios.Load
+import HIE.Bios.Ghc.Api
+import HIE.Bios.Ghc.Load
